Make AIE PSP code common for aie2 and aie4
